commit to main

+++ api/database.py
@@ -13,6 +13,7 @@ posts_collection = social_media_db.get_collection("posts")
 def follow_user(user_id, follow_id):
 	if user_id != follow_id:
 		user_data = users_collection.find_one({'_id': user_id}) 
+		print("user_data found")
 		followings = user_data['following']
 		
 		user_data = users_collection.find_one({'_id': follow_id})
@@ -26,19 +27,19 @@ def follow_user(user_id, follow_id):
 	return "Follow failed : follow_id same as user_id"
 
 def unfollow_user(user_id, unfollow_id):
-	if user_id != follow_id:
+	if user_id != unfollow_id:
 		user_data = users_collection.find_one({'_id': user_id}) 
 		followings = user_data['following']
 
-		if follow_id in followings:
-			users_collection.update_one({'_id': user_id}, {'$pull':{'following': follow_id}})
-			users_collection.update_one({'_id': follow_id}, {'$pull':{'follower': user_id}})
+		if unfollow_id in followings:
+			users_collection.update_one({'_id': user_id}, {'$pull':{'following': unfollow_id}})
+			users_collection.update_one({'_id': unfollow_id}, {'$pull':{'follower': user_id}})
 			return "OK"
-		return "Unfollow failed : follow_id not in following list of user_id"
-	return "Unfollow failed : follow_id same as user_id"
+		return "Unfollow failed : unfollow_id not in following list of user_id"
+	return "Unfollow failed : unfollow_id same as user_id"
 
-def get_user(email):
-	user_data = users_collection.find_one({'_id': email})
+def get_user(user_id):
+	user_data = users_collection.find_one({'_id': user_id})
 	if user_data:
 		return User(user_data['_id'], user_data['password'], user_data['name'], user_data['following'], user_data['follower'], user_data['liked_posts']) 
 	return None
@@ -46,33 +47,33 @@ def get_user(email):
 
 def add_post(post):
 	if post.title and post.desc:
-		posts_collection.insert_one({'_id': post.post_id, 'created_by': post.created_by, 'title': post.title, 'desc': post.desc, 'created_at': post.created_at})
-		return "OK"
-	else:
-		return "Title or Desc not found"
+		posts_collection.insert_one({'_id': post.post_id, 'created_by': post.created_by, 'title': post.title, 'desc': post.desc, 'created_at': post.created_at, 'comments': post.comments, 'likes_count': post.likes_count})
+		return post
+	return None
 
-def delete_post(post_id):
+def delete_post(post_id, user_id):
 	post = posts_collection.find_one({'_id': post_id})
-	if post:
+	if post and post['created_by'] == user_id:
 		posts_collection.delete_one({'_id': post_id})
 		return "OK"
 	return "Post not found"
 
 
-def like_post(user_id, post_id):
+def like_post(post_id, user_id):
 	post = posts_collection.find_one({'_id': post_id})
+	print(post)
 	if post:
 		user_data = users_collection.find_one({'_id': user_id})
 		liked_posts_list = user_data['liked_posts']
 		if post_id not in liked_posts_list:
-			users_collection.update_one({'_id': user_id}, {'$addToSet': { 'liked_posts': post_id }})
+			users_collection.update_one({'_id': user_id}, {'$addToSet': {'liked_posts': post_id }})
 			posts_collection.update_one({'_id': post_id}, {'$inc': {'likes_count': 1}})
 			return "OK"
 		else:
 			return "Post already in liked posts list of user_id"
 	return "Post ID not found"
 
-def unlike_post(user_id, id):
+def unlike_post(user_id, post_id):
 	post = posts_collection.find_one({'_id': post_id})
 	if post:
 		user_data = users_collection.find_one({'_id': user_id})
@@ -97,5 +98,5 @@ def add_comment(user_id, post_id, comment):
 def get_post(post_id):
 	post = posts_collection.find_one({'_id': post_id})
 	if post:
-		return Post(post['post_id'], post['created_by'], post['title'], post['desc'], post['created_at'], post['comment'], post['likes_count'])
-	return "Post ID not found"
+		return Post(post['_id'], post['created_by'], post['title'], post['desc'], post['created_at'], post['comments'], post['likes_count'])
+	return None
